Revert "cleanup(compact_level0_phase1): some commentary and wrapping into block expressions (#8544)"
This reverts commit d95b46f3f35eaae3aaec4f218986ece33acab052.
